rename package process to manager
