+++ pkg/manager/daemon_adaptor.go
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package process
+package manager
 
 import (
 	"os"
