+++ pkg/manager/manager.go
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package process
+package manager
 
 import (
 	"context"
