+++ pkg/manager/monitor_linux.go
@@ -17,7 +17,7 @@
    limitations under the License.
 */
 
-package process
+package manager
 
 import (
 	"net"
