+++ pkg/manager/monitor_test.go
@@ -17,7 +17,7 @@
    limitations under the License.
 */
 
-package process
+package manager
 
 import (
 	"context"
